Refactorizando codigo

+++ database/migrations/2022_12_24_040926_create_personas_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('cedula')->unique();
+            $table->bigInteger('cedula')->unique();
             $table->string('apellidos');
             $table->string('nombres');
             $table->string('email')->unique();
